chore: alot of code readability improve+get ready to merge
