--- src/components/item-display/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ItemDisplay';
