+++ src/components/wynncraft/item/StarRating.tsx
@@ -1,15 +1,16 @@
 "use client";
 
+import { getStarsFromRollPercentage } from '@/lib/itemUtils';
 import React from 'react';
 
 interface StarRatingProps {
-  stars: number;
+  percentage: number;
 }
 
-const StarRating: React.FC<StarRatingProps> = ({ stars }) => {
+const StarRating: React.FC<StarRatingProps> = ({ percentage }) => {
   return (
     <span className="inline-flex items-center">
-      {Array.from({ length: stars }).map((_, i) => (
+      {Array.from({ length: getStarsFromRollPercentage(percentage) }).map((_, i) => (
         <span
           key={i}
           className="text-xs"
@@ -21,4 +22,4 @@ const StarRating: React.FC<StarRatingProps> = ({ stars }) => {
   );
 };
 
-export default StarRating;
+export { StarRating };
