+++ src/map/itemMap.tsx
@@ -1,0 +1,39 @@
+export const colorMap: Record<string, string> = {
+  baseHealth: 'text-[#AA0000]',
+  baseDamage: 'text-[#FFAA00]',
+  Earth: 'text-[#00AA00]',
+  Thunder: 'text-[#FFFF55]',
+  Water: 'text-[#55FFFF]',
+  Fire: 'text-[#FF5555]',
+  Air: 'text-[#FFFFFF]',
+};
+
+export const textMap: Record<string, string> = {
+  baseHealth: 'Health',
+  baseDamage: 'Neutral',
+  Earth: 'Earth',
+  Thunder: 'Thunder',
+  Water: 'Water',
+  Fire: 'Fire',
+  Air: 'Air',
+};
+
+export const POWDER_ELEMENTS: Record<number, string> = {
+  1: "Earth",
+  2: "Thunder", 
+  3: "Water",
+  4: "Fire", 
+  5: "Air", 
+};
+
+
+export const RARITY_COLORS = {
+  normal: "#FFFFFF",
+  unique: "#FFFF55",
+  rare: "#55FFFF",
+  legendary: "#AA00AA",
+  fabled: "#FF55FF",
+  mythic: "#AA00FF",
+  set: "#55FF55",
+  crafted: "#A0A0A0",
+};
